test: modify LottoTest to move calculateTotalPrize
